# Backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_odatlar-bot_backend_internal_dto.CreateGuideCourseRequest:
    properties:
      children:
        items:
          type: integer
        type: array
      description:
        maxLength: 5000
        type: string
      is_published:
        type: boolean
      slug:
        maxLength: 100
        minLength: 1
        type: string
      sort_order:
        type: integer
      title:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - children
    - slug
    - title
    type: object
...
  github_com_odatlar-bot_backend_internal_dto.GuideCourseListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.GuideCourseResponse'
        type: array
    type: object
  github_com_odatlar-bot_backend_internal_dto.GuideCourseResponse:
    properties:
      children:
        items:
          type: integer
        type: array
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      isPublished:
        type: boolean
      lessonCount:
        type: integer
      sectionCount:
        type: integer
      sortOrder:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  github_com_odatlar-bot_backend_internal_dto.UpdateGuideCourseRequest:
    properties:
      children:
        items:
          type: integer
        type: array
      description:
        maxLength: 5000
        type: string
      is_published:
        type: boolean
      slug:
        maxLength: 100
        minLength: 1
        type: string
      sort_order:
        type: integer
      title:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - children
    - slug
    - title
    type: object
...
  github_com_odatlar-bot_backend_pkg_upload.SavedFile:
    properties:
      ext:
        type: string
      path:
        type: string
      sizeBytes:
        type: integer
      sizeLabel:
        type: string
      url:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@odatlar-bot.uz
    name: API Support
  description: Admin panel, Bot Runtime (Telegram WebApp), Qo'llanma (videolar, kurslar,
    fayllar) va Bildirishnomalar APIlari.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Odatlar Bot API
  version: "1.0"
paths:
...
  /bot/guides/courses:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.GuideCourseListResponse'
      security:
      - BearerAuth: []
      summary: Qo'llanma kurslari ro'yxati (admin)
      tags:
      - Guide Courses
    post:
      consumes:
      - application/json
      parameters:
      - description: Kurs
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.CreateGuideCourseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.GuideCourseResponse'
      security:
      - BearerAuth: []
      summary: Yangi qo'llanma kursi
      tags:
      - Guide Courses
  /bot/guides/courses/{id}:
    delete:
      parameters:
      - description: Kurs slug yoki ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Qo'llanma kursini o'chirish
      tags:
      - Guide Courses
    get:
      parameters:
      - description: Kurs slug yoki ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.GuideCourseResponse'
      security:
      - BearerAuth: []
      summary: Bitta qo'llanma kursi (admin)
      tags:
      - Guide Courses
    put:
      consumes:
      - application/json
      parameters:
      - description: Kurs slug yoki ID
        in: path
        name: id
        required: true
        type: string
      - description: Kurs
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.UpdateGuideCourseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_internal_dto.GuideCourseResponse'
      security:
      - BearerAuth: []
      summary: Qo'llanma kursini yangilash
      tags:
      - Guide Courses
...
  /bot/guides/upload/file:
    post:
      consumes:
      - multipart/form-data
      description: Hujjat fayl (pdf, txt, doc, docx, zip va boshqalar)
      parameters:
      - description: Fayl
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_upload.SavedFile'
      security:
      - BearerAuth: []
      summary: Qo'llanma fayli yuklash
      tags:
      - Guide Files
  /bot/guides/upload/poster:
    post:
      consumes:
      - multipart/form-data
      description: Faqat rasm fayli (jpg, png, webp, gif). JSON orqali URL qabul qilinmaydi.
      parameters:
      - description: Poster rasm
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_upload.SavedFile'
      security:
      - BearerAuth: []
      summary: Poster rasm yuklash
      tags:
      - Guide Videos
  /bot/guides/upload/video:
    post:
      consumes:
      - multipart/form-data
      description: Video fayl (mp4, webm, mov). Tashqi URL o'rniga yuklangan path
        ishlatiladi.
      parameters:
      - description: Video fayl
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_odatlar-bot_backend_pkg_upload.SavedFile'
      security:
      - BearerAuth: []
      summary: Video fayl yuklash
      tags:
      - Guide Videos
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token. Format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
  BotToken:
    description: 'Telegram bot token. Format: Bearer {bot_token}'
    in: header
    name: Authorization
    type: apiKey
  TelegramInitData:
    description: 'Telegram WebApp initData. Alternativa: Authorization: tma {initData}'
    in: header
    name: X-Telegram-Init-Data
    type: apiKey
swagger: "2.0"